requirements.txt
requests==2.31.0
PyExecJs==1.5.1
urllib3==1.26.15
